Some issues occurred due to moving of files, solved them
