+++ client/src/components/LoginRegister/Login.jsx
@@ -1,7 +1,6 @@
-import React, { useState, makeSyles } from "react";
+import React, { useState } from "react";
 import { FormControl, Input, InputLabel, Button } from "@material-ui/core";
 import "./css/FormStyle.css";
-
 const Login = () => {
   const [email, setEmail] = useState("");
   const [number, setNumber] = useState("");
